requirements.txt
streamlit
python-dotenv
langchain
langchain_community
langchain-groq
faiss-cpu
sentence-transformers

pypdf
PyPDF2
